+++ azext_edge/edge/providers/orchestration/upgrade2.py
@@ -22,6 +22,7 @@ from rich.progress import (
 )
 from rich.table import Table, box
 
+from ....constants import DEFAULT_REGISTRY_HOST
 from ...util import parse_kvp_nargs, should_continue_prompt
 from ...util.machinery import scoped_semver_import
 from .common import (
@@ -48,7 +49,6 @@ logger = get_logger(__name__)
 console = Console()
 
 
-DEFAULT_REGISTRY_HOST = "mcr.microsoft.com"
 IOT_OPS_DELAY = 30  # seconds
